# server/openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "API for adding, retrieving and removing centers and events",
    "version": "1",
    "title": "EventsManager API",
    "contact": {
      "email": "adegold71@gmail.com"
    }
  },
  "basePath": "/api/v1/",
...
  "schemes": [
    "https", "http"
  ],
  "paths": {
...
    "/events": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Create new user event",
        "operationId": "EventController.createEvent",
        "consumes" :[
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "userSecurity": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Details used to create new user event",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SubmitEvent"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful request",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Authentication error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "User/Center not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "409": {
            "description": "Date conflict",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Fetches existing user events",
        "operationId": "EventController.fetchUserEvents",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "userSecurity": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "schema": {
              "items": {
                "$ref": "#/definitions/Event"
              }
            }
          }
        }
      }
    },
    "/events/{id}": {
      "put": {
        "tags": [
          "Events"
        ],
        "summary": "Modifies existing user event",
        "operationId": "EventController.modifyEvent",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "security": [
          {
            "userSecurity": []
          }
        ],
        "parameters": [
          { 
            "name": "id",
            "in": "path",
            "description": "id of the event to be modified",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Details used to create new user event",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SubmitEvent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "schema": {
              "items": {
                "$ref": "#/definitions/Event"
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Authentication error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "User/Center not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "409": {
            "description": "Date conflict",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "summary": "Deletes existing user event",
        "operationId": "EventController.deleteEvent",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "userSecurity": []
          }
        ],
        "parameters": [
          { 
            "name": "id",
            "in": "path",
            "description": "id of the event to be deleted",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "schema": {
              "items": {
                "$ref": "#/definitions/Success"
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Authentication error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "User/Event not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/events/{id}/decline": {
      "put": {
        "tags": [
          "Events"
        ],
        "summary": "Declines existing user event",
        "operationId": "EventController.declineUserEvent",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "adminSecurity": []
          }
        ],
        "parameters": [
          { 
            "name": "id",
            "in": "path",
            "description": "id of the event to be declined",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "schema": {
              "items": {
                "$ref": "#/definitions/Success"
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Authentication error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "User/Event not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "userSecurity": {
      "type": "apiKey",
      "name": "token",
      "in": "query"
    },
    "adminSecurity": {
      "type": "apiKey",
      "name": "token",
      "in": "query"
    }
  },
  "definitions": {
    "SubmitEvent": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "centerId": {
          "type": "integer"
        },
        "start": {
          "type": "string"
        },
        "guests": {
          "type": "integer"
        },
        "days": {
          "type": "integer"
        }
      },
      "required": ["name", "type", "centerId", "start", "days", "guests"]
    },
    "Event": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "event": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "center": {
              "type": "string"
            },
            "start": {
              "type": "string"
            },
            "end": {
              "type": "string"
            },
            "id": {
              "type": "integer"
            },
            "guests": {
              "type": "integer"
            },
            "days": {
              "type": "integer"
            }
          },
          "required": ["name", "type", "center", "start", "days", "guests", "id"]
        }
      },
      "required": ["status", "event"]
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      },
      "required": [
        "status", "error"
      ]
    },
    "Success": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        }
      },
      "required": ["status"]
    }
  }
}